spi-xilinux-ps: Fix off by one bug.
When determining the baud rate divisor, setting the value to 0b000
results in the ps7 spi bus not working correctly.

The loop to determine the baud rate divisor terminated with the value of
8 which after masking results in a value of 0 being set for the divisori
(which is invalid).  Terminating with a value of 7 allows the divisor to
be set to a valid value.