Pull in API's needed for sound/soc/adi drivers.
As part of the adau1761 backport, the adau1761 files in the
sound/soc/adi directory use APIs not provided in the 3.8 kernel. This
commit pulls enough from the
https://github.com/analogdevicesinc/linux/commits/audio_zynq branch to
allow everything to compile.

Kernel boots to login prompt, but alsa still doesn't see the adau1761
sound card. More tweaking of the DTS is probably needed.